Introduced roles support. You can now set a role in the conf file. If it is non-empty then a role with a matching string must be preset in the users list of roles before the bot will respond to any commands. (#19)
One thing to note is that by adding this support the bot will no longer work across multiple guilds/servers. If that is something that is needed then role support should not be used. A lot more work would have to be done to support multiple server roles if necessary.